requirements.txt
faster-whisper>=1.0.0
pytest>=8.0.0
openai>=1.0.0
python-dotenv>=1.0.0
presidio-analyzer>=2.2.35
presidio-anonymizer>=2.2.35
spacy>=3.7.0

# --- Phase 2: FastAPI Web Server & Database ---
fastapi[standard]>=0.115.0
uvicorn>=0.30.0
sqlmodel>=0.0.22

# --- Phase 3: Security & Upload ---
pyjwt>=2.8.0
passlib[bcrypt]>=1.7.4
bcrypt==3.2.2
python-multipart>=0.0.9
